+++ base.ts
@@ -26,6 +26,7 @@ export interface Config {
 	readonly namespaces: Record<string, string>;
 	readonly nsid: Record<string, number>;
 	readonly variable: string[];
+	readonly functionHook: string[];
 	readonly parserFunction: [Record<string, string>, Record<string, string> | string[], string[], string[]];
 	readonly doubleUnderscore: [string[], string[], Record<string, string>?, Record<string, string>?];
 	readonly protocol: string;